requirements.txt
# Core dependencies
pydantic>=2.0
pydantic-settings>=2.0
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Async and networking
aiohttp>=3.9.0
asyncio

# Data processing and ML
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# LLM providers (install the ones you need)
openai>=1.0.0
anthropic>=0.18.0

# Optional but recommended
pytest>=7.0.0
pytest-asyncio>=0.21.0
